Code cleanup, still WIP
